добавлена обработка исключений, отредактирован контроллер объявлений

+++ src/Board/Infrastructure/Board.Infrastructure.DataAccess/Contexts/Posts/Repositories/PostRepository.cs
@@ -4,13 +4,13 @@ using Board.Domain.Categories;
 using Board.Domain.Comments;
 using Board.Domain.Posts;
 using Board.Infrastructure.Repository;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static Board.Infrastructure.DataAccess.Contexts.Posts.Repositories.PostRepository;
+using System.Security.Claims;
+
 
 namespace Board.Infrastructure.DataAccess.Contexts.Posts.Repositories;
 
@@ -18,11 +18,13 @@ namespace Board.Infrastructure.DataAccess.Contexts.Posts.Repositories;
 public class PostRepository : IPostRepository
 {
     private readonly IRepository<Post> _repository;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public PostRepository(IRepository<Post> postRepository)
+    public PostRepository(IRepository<Post> postRepository, IHttpContextAccessor httpContextAccessor)
     {
         _repository = postRepository;
-    }   
+        _httpContextAccessor = httpContextAccessor;
+    }
 
     /// <inheritdoc/>
     public async Task<Guid> AddPostAsync(Post dto, CancellationToken cancellationToken)
@@ -34,6 +36,18 @@ public class PostRepository : IPostRepository
     /// <inheritdoc/>
     public async Task DeleteById(Guid id, CancellationToken cancellationToken)
     {
+        //получение идентификатора пользователя из контекста 
+        var claims = _httpContextAccessor.HttpContext.User.Claims;
+        var claimId = claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+        if (string.IsNullOrWhiteSpace(claimId)) throw new Exception("ClaimId is null!!!");
+        var UserId = Guid.Parse(claimId);
+
+        //Получение доменной модели поста и обработка исключений
+        var entity = await _repository.GetByIdAsync(id, cancellationToken);
+        if (entity == null) throw new Exception("Введеный идентификатор не принадлежит ни одному существующему объявлению!");
+        if (entity.AccountId != id) throw new Exception("Текущий пользователь не может удалить пост другого пользователя.");
+
+
         await _repository.DeleteByIdAsync(id, cancellationToken);
     }
 
@@ -59,6 +73,18 @@ public class PostRepository : IPostRepository
     /// <inheritdoc/>
     public async Task<Post> UpdateAsync(Guid id, Post dto, CancellationToken cancellationToken)
     {
+        //получение идентификатора пользователя из контекста 
+        var claims = _httpContextAccessor.HttpContext.User.Claims;
+        var claimId = claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+        if (string.IsNullOrWhiteSpace(claimId)) throw new Exception("ClaimId is null!!!");
+        var UserId = Guid.Parse(claimId);
+
+        //Получение доменной модели поста и обработка исключений
+        var entity = await _repository.GetByIdAsync(id, cancellationToken);
+        if (entity == null) throw new Exception("Введеный идентификатор не принадлежит ни одному существующему объявлению!");
+        if (entity.AccountId != id) throw new Exception("Текущий пользователь не может редактировать пост другого пользователя.");
+
+
         var model = await _repository.UpdateAsync(id, dto, cancellationToken);
         return model;
     }
